Reorganise files location in QPack to better readability
