+++ src/http/qpack/primitive/mod.rs
@@ -1,0 +1,13 @@
+// This is only here because qpack is new and quinn no uses it yet.
+// TODO remove allow dead code
+#![allow(unused_imports)]
+
+pub mod iocontext;
+pub use self::iocontext::StarterByte;
+
+pub mod parser;
+pub use self::parser::Parser;
+
+pub mod dump;
+pub use self::dump::{Dump, StringEncoding};
+
