+++ src/http/qpack/table/mod.rs
@@ -1,0 +1,12 @@
+// This is only here because qpack is new and quinn no uses it yet.
+// TODO remove allow dead code
+#![allow(unused_imports)]
+
+pub mod field;
+pub use self::field::HeaderField;
+
+pub mod dynamic;
+pub use self::dynamic::{DynamicTable, ErrorKind};
+
+pub mod static_;
+pub use self::static_::StaticTable;
